Fixed up the defend command. No issues detected.

+++ CommandEverything/CommandEverythingCPP/CommandDefend.cpp
@@ -135,20 +135,32 @@ void CommandDefend::Defend()
 				string converted(ws.begin(), ws.end());
 				//https://stackoverflow.com/questions/45288937/avoid-killing-a-specific-process-c?noredirect=1#comment77540465_45288937
 
-				//Kill process.
-				if (TerminateProcess(hProcess, 1))
+				char drive[4];
+				char directory[MAX_PATH];
+				char filename[MAX_PATH];
+				char extension[MAX_PATH];
+				_splitpath(converted.c_str(), drive, directory, filename, extension);
+
+				if (filename[0] != 'c' || filename[1] != 'o' || filename[2] != 'n' || filename[3] != 'h'
+					|| filename[4] != 'o' || filename[5] != 's' || filename[6] != 't'
+					|| extension[0] != '.' || extension[1] != 'e' || extension[2] != 'x'
+					|| extension[3] != 'e')
 				{
-					string msg = "Killed a process: ";
-					msg.append(converted);
-					Console->WriteLine(&msg);
-				}
-				else
-				{
-					this->FailedHitList.push_back(aProcesses[i]);
-					string msg = "Failed to kill a process: ";
-					msg.append(converted);
-					Console->WriteLine(&msg);
-					Console->WriteLine(&to_string(GetLastError()));
+					//Kill process.
+					if (TerminateProcess(hProcess, 1))
+					{
+						string msg = "Killed a process: ";
+						msg.append(converted);
+						Console->WriteLine(&msg);
+					}
+					else
+					{
+						this->FailedHitList.push_back(aProcesses[i]);
+						string msg = "Failed to kill a process: ";
+						msg.append(converted);
+						Console->WriteLine(&msg);
+						Console->WriteLine(&to_string(GetLastError()));
+					}
 				}
 			}
 		}
